+++ cmn-data-svc-client/src/test/java/org/acumos/cds/client/test/ClientMethodTest.java
@@ -1069,6 +1069,16 @@ public class ClientMethodTest {
 			logger.info("Client failed as expected: {}", ex.toString());
 		}
 		try {
+			client.getPeerAccessCatalogIds("peerId");
+		} catch (ResourceAccessException ex) {
+			logger.info("Client failed as expected: {}", ex.toString());
+		}
+		try {
+			client.getCatalogAccessPeers("catalogId");
+		} catch (ResourceAccessException ex) {
+			logger.info("Client failed as expected: {}", ex.toString());
+		}
+		try {
 			client.getRightToUse(0L);
 		} catch (ResourceAccessException ex) {
 			logger.info("Client failed as expected: {}", ex.toString());