+++ docs/release-notes.rst
@@ -24,18 +24,24 @@ The client and server are released together.  The client is available
 as a jar file in the Acumos/LF Nexus repository. The server is
 available as a Docker image in the Acumos/LF Docker registry.
 
+Version 2.2.6, 1 August 2019
+----------------------------
+* Add query method to get peers for a restricted catalog (`ACUMOS-3298 <https://jira.acumos.org/browse/ACUMOS-3298>`_)
+* Requires database schema version 2.2
+
 Version 2.2.5, 16 July 2019
 ---------------------------
 * Add headers to discover texts (`ACUMOS-2992 <https://jira.acumos.org/browse/ACUMOS-2992>`_)
 * Add kernel type CC (`ACUMOS-3147 <https://jira.acumos.org/browse/ACUMOS-3147>`_)
 * Add kernel type JS (`ACUMOS-3148 <https://jira.acumos.org/browse/ACUMOS-3148>`_)
 * Update default catalog names (`ACUMOS-3209 <https://jira.acumos.org/browse/ACUMOS-3209>`_)
+* Requires database schema version 2.2
 
 Version 2.2.4, 23 May 2019
 --------------------------
 * Restore automatic modified-date updates (`ACUMOS-2949 <https://jira.acumos.org/browse/ACUMOS-2949>`_)
-* Requires database schema version 2.2
 * Extend migration script for solutions with mixed access types (`ACUMOS-3110 <https://jira.acumos.org/browse/ACUMOS-3110>`_)
+* Requires database schema version 2.2
 
 Version 2.2.3, 15 May 2019
 --------------------------